Move plugin assets out to an interface
