+++ src/main/java/com/avairebot/commands/system/plugin/InstallPlugin.java
@@ -26,11 +26,11 @@ import com.avairebot.commands.CommandMessage;
 import com.avairebot.commands.system.PluginCommand;
 import com.avairebot.contracts.commands.plugin.PluginSubCommand;
 import com.avairebot.contracts.plugin.Plugin;
+import com.avairebot.contracts.plugin.PluginAsset;
 import com.avairebot.contracts.plugin.PluginRelease;
 import com.avairebot.contracts.plugin.PluginSourceManager;
 import com.avairebot.exceptions.InvalidPluginException;
 import com.avairebot.exceptions.InvalidPluginsPathException;
-import com.avairebot.plugin.PluginAsset;
 import com.avairebot.plugin.PluginLoader;
 
 import java.io.BufferedInputStream;
