+++ src/main/java/com/avairebot/contracts/commands/plugin/PluginSubCommand.java
@@ -26,8 +26,8 @@ import com.avairebot.Constants;
 import com.avairebot.commands.CommandMessage;
 import com.avairebot.commands.system.PluginCommand;
 import com.avairebot.contracts.plugin.Plugin;
+import com.avairebot.contracts.plugin.PluginAsset;
 import com.avairebot.contracts.plugin.PluginRelease;
-import com.avairebot.plugin.PluginAsset;
 import com.avairebot.plugin.PluginHolder;
 import com.avairebot.plugin.PluginLoader;
 import com.avairebot.plugin.translators.PluginHolderTranslator;
